requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
torch==2.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
chromadb==0.4.18
pydantic==2.5.2
python-multipart
python-dotenv
scikit-learn
pandas
psycopg2-binary